Move styles from `styles/card-styles.js` to `styles.js`
