+++ packages/components/src/card/divider.js
@@ -6,7 +6,7 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import { DividerUI } from './styles/card-styles';
+import { DividerUI } from './styles';
 
 export function CardDivider( props ) {
 	const { className, ...additionalProps } = props;
